% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi.r
\docType{data}
\name{.opi_env}
\alias{.opi_env}
\title{Global environment for OPI to hold machine specific constants, etc.}
\format{
An object of class \code{environment} of length 11.
}
\usage{
.opi_env
}
\description{
Global environment for OPI to hold machine specific constants, etc.
}
\keyword{datasets}
