% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi.r
\docType{data}
\name{.OpiEnv}
\alias{.OpiEnv}
\title{Environment holding the state of the OPI}
\format{
An object of class \code{environment} of length 13.
}
\usage{
.OpiEnv
}
\description{
Holds the chosen machine and any parameters forming the state of the machine.
}
\keyword{datasets}
