% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KowaAP7000Client.r
\name{kowa.presentTemporal}
\alias{kowa.presentTemporal}
\title{Present temporal stim on Kowa AP7000 (internal use)}
\description{
Implementation of opiPresent for the Kowa AP7000 machine.
Version for opiKineticStimulus.

This is for internal use only. Use \code{\link[=opiPresent]{opiPresent()}} with
\code{stim} as class \code{opiStaticStimulus} and you will get the Value back.
}
