% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi.r
\name{open_socket}
\alias{open_socket}
\title{Open a socket on ip and port.}
\usage{
open_socket(ip, port, machineName)
}
\arguments{
\item{ip}{IP address of socket}

\item{port}{TCP port of socket}

\item{machineName}{Machine name for error message}
}
\value{
Socket or NULL on error
}
\description{
Internal use only.
}
