\name{OPI-package}
\alias{OPI-package}
\alias{OPI}
\docType{package}
\title{
Open Perimetry Interface
}
\description{

This package contains the current version of the Open Perimetry
Interface (OPI), for controlling visual field machines using R. The
OPI is an R standard for interfacing with visual field testing
machines (perimeters). It specifies basic functions that allow many
visual field tests to be constructed.  As of June 2012 it is partially
implemented on the Octopus 900 and the Heidelberg Edge Perimeter.
It also has a cousin: the R package VisualFields, which has tools
for analysing and manipulating visual field data.

}
\details{
    \tabular{ll}{
     Package: \tab OPI\cr
     Type: \tab Package\cr
     Version: \tab 1.0\cr
     Date: \tab 2012-06-19\cr
     License: \tab GPL-3\cr
    }
}
\author{
Andrew Turpin <aturpin@unimelb.edu.au>
Maintainer: Andrew Turpin <aturpin@unimelb.edu.au>
}
\references{
http://perimetry.org/OPI
}
\keyword{ package }
\seealso{
visualFields
}
