\name{opi.procedure.load}
\alias{opi.procedure.load}
\title{
opi.procedure.load
}
\description{
Load procedure state that is saved with \code{opi.procedure.save}.
}
\usage{
opi.procedure.load(filename)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
\item{filename}{
    Name of file from which to load.
}
}
\details{
Load procedure state that is saved with \code{opi.procedure.save}.
}
\value{
Returns a list of states.
}

\references{
Please cite:
A. Turpin, P.H. Artes and A.M. McKendrick
"The Open Perimetry Interface: An enabling tool for clinical visual psychophysics",
Journal of Vision 12(11) 2012.

http://perimetry.org/OPI
}

\author{
Andrew Turpin <aturpin@unimelb.edu.au>
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
\code{\link{ZEST}}, \code{\link{FT}}, \code{\link{opi.procedure.save}}
}

% Add one or more standard keywords, see file 'KEYWORDS' in the
% R documentation directory.
\keyword{ misc }
