% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.r
\name{opiClose_for_Display}
\alias{opiClose_for_Display}
\title{Implementation of opiClose for the Display machine.}
\value{
A list containing:
\itemize{
\item \code{err} \code{NULL} if there was no error, a string message if there is an error.
}
}
\description{
This is for internal use only. Use \code{\link[=opiClose]{opiClose()}} after
\code{chooseOPI("Display")} to call this function.
}
\examples{
chooseOpi("Display")
opiInitialise(list(port = 50001, ip = "localhost"))
opiSetup(list(eye = "BOTH"))
result <- opiClose()

}
\seealso{
\code{\link[=opiClose]{opiClose()}}
}
