% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Envision.r
\name{opiClose_for_Envision}
\alias{opiClose_for_Envision}
\title{Implementation of opiClose for the Envision HMP.}
\value{
A list containing:
\itemize{
\item \code{err} \code{NULL} if there was no error, a string message if there is an error.
}
}
\description{
This is for internal use only. Use \code{\link[=opiClose]{opiClose()}} after
\code{chooseOPI("Envision")} to call this function.
}
\examples{
\dontrun{
chooseOpi("Envision")
opiInitialise(address = list(port = 50001, ip = "localhost"))
opiSetup(list(eye = "BOTH"))
result <- opiClose()
}

}
\seealso{
\code{\link[=opiClose]{opiClose()}}
}
