% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KowaAP7000Client.r
\name{opiClose_for_KowaAP7000}
\alias{opiClose_for_KowaAP7000}
\title{Implementation of opiClose for the Kowa AP7000 machine.}
\value{
Returns \code{list(err = NULL)}.
}
\description{
This is for internal use only. Use \code{\link[=opiClose]{opiClose()}} with the same parameters.
}
