% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIA.r
\name{opiClose_for_MAIA}
\alias{opiClose_for_MAIA}
\title{Implementation of opiClose for the MAIA machine.}
\value{
A list with elements
\itemize{
\item \code{err}, which is an error code, NULL for no error
\item \code{fixations}, which is a matrix one row per fixation and four columns:
\itemize{
\item \code{time} (same as \code{time_hw} in \code{opiPresent})
\item \code{flags} Bit 0(same as \code{time_hw} in \code{opiPresent})
\itemize{
\item bit 0: 1 if the tracking event is valid (quality>threshold), 0 otherwise
(in this case, x and y shall be =0)
\item bit 2: 1 if a stimulus was being presented during the tracking event, 0 otherwise
\item bit 3: 1 if the tracking was enabled during the tracking event, 0 otherwise
\item bits 16..31: ID of the stimulus being presented (if bit 2 == 1)
}
\item \code{x} (degrees relative to the centre of the image returned by \code{opiInitialise} - not the PRL)
\item \code{y} (as for x)
}
}
}
\description{
This is for internal use only. Use \code{\link[=opiClose]{opiClose()}} with
these Arguments and you will get the Value back.
}
