% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Octopus600.r
\name{opiClose_for_O600}
\alias{opiClose_for_O600}
\title{Implementation of opiClose for the O600 machine.}
\value{
Returns \code{list(err = NULL)}.
}
\description{
This is for internal use only. Use \code{\link[=opiClose]{opiClose()}} with the same parameters.
}
