% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Octopus900.r
\name{opiClose_for_Octopus900}
\alias{opiClose_for_Octopus900}
\title{Implementation of opiClose for the Octopus900 machine.}
\value{
Returns \code{list(err = NULL)}.
}
\description{
This is for internal use only. Use \code{\link[=opiClose]{opiClose()}} with the same parameters.
}
