% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimHensonRT.r
\name{opiClose_for_SimHensonRT}
\alias{opiClose_for_SimHensonRT}
\title{opiClose_for_SimHensonRT}
\value{
A list with elements:
\itemize{
\item \code{err} Always \code{NULL}.
}
}
\description{
Does nothing.
}
