% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNo.r
\name{opiClose_for_SimNo}
\alias{opiClose_for_SimNo}
\title{opiClose_for_SimNo}
\value{
A list with elements:
\itemize{
\item \code{error} Always \code{FALSE}.
\item \code{msg} A string "Close OK".
}
}
\description{
Does nothing.
}
