% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi.r
\docType{data}
\name{opi.implementations}
\alias{opi.implementations}
\alias{opiDistributor}
\title{FOR INTERNAL USE ONLY}
\format{
\code{opi.implementations} is a \code{list} containing a catalog of
all specific methods that are dependent on the OPI implementation selected
with chooseOpi
}
\usage{
opi.implementations

opiDistributor(operation, ...)
}
\arguments{
\item{operation}{A general OPI operation of the following methods to: \code{opiInitialize},
\code{opiPresent} \code{opiClose}, \code{opiSetBackground},
\code{opiQueryDevice}}

\item{...}{other parameters to pass to the methods}
}
\description{
The method \code{opiDistributor} searches for the specific
method of a general OPI \code{operation}, which depends on the OPI
implementation selected with \code{\link{chooseOpi}}. It returns an error if no
OPI implementation has been selected yet. A catalog of all specific methods are
listed in \code{opi.implementations}.
}
\keyword{datasets}
