% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Display.r
\name{opiInitialise_for_Display}
\alias{opiInitialise_for_Display}
\title{Implementation of opiInitialise for the Display machine.}
\arguments{
\item{address}{A list containing:
\itemize{
\item \code{port} TCP port of the OPI Monitor.
\item \code{ip} IP Address of the OPI Monitor.
}}
}
\value{
A list containing:
\itemize{
\item \code{err} \code{NULL} if there was no error, a string message if there is an error.
}
}
\description{
This is for internal use only. Use \code{\link[=opiInitialise]{opiInitialise()}} after
\code{chooseOPI("Display")} to call this function.
}
\details{
\code{port} can take on values in the range \code{[0, 65535]}.
}
\examples{
chooseOpi("Display")
result <- opiInitialise(address = list(port = 50001, ip = "localhost"))

}
\seealso{
\code{\link[=opiInitialise]{opiInitialise()}}
}
