% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Envision.r
\name{opiInitialise_for_Envision}
\alias{opiInitialise_for_Envision}
\title{Implementation of opiInitialise for the Envision HMP.}
\arguments{
\item{address}{A list containing:
\itemize{
\item \code{port} TCP port of the Envision's OPI.
\item \code{ip} IP Address of the Envision's OPI.
}}
}
\value{
A list containing:
\itemize{
\item \code{err} \code{NULL} if there was no error, a string message if there is an error.
}
}
\description{
This is for internal use only. Use \code{\link[=opiInitialise]{opiInitialise()}} after
\code{chooseOPI("Envision")} to call this function.
}
\details{
\code{port} can take on values in the range \code{[0, 65535]}.
}
\examples{
\dontrun{
chooseOpi("Envision")
result <- opiInitialise(address = list(port = 50001, ip = "localhost"))
}

}
\seealso{
\code{\link[=opiInitialise]{opiInitialise()}}
}
