% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNo.r
\name{opiInitialise_for_SimNo}
\alias{opiInitialise_for_SimNo}
\title{opiInitialise_for_SimNo}
\arguments{
\item{...}{Any object you like, it is ignored.}
}
\value{
A list with elements:
\itemize{
\item \code{err} Always \code{NULL}.
}
}
\description{
Does nothing.
}
