% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi.r
\name{opiKineticStimulus}
\alias{opiKineticStimulus}
\title{For backwards compatibility. Used by Octopus900 and KowaAP7000.}
\usage{
opiKineticStimulus()
}
\description{
For backwards compatibility. Used by Octopus900 and KowaAP7000.
}
