% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Envision.r
\name{opiPresent_for_Envision}
\alias{opiPresent_for_Envision}
\title{Implementation of opiPresent for the Envision HMP.}
\arguments{
\item{stim}{A list containing:
\itemize{
\item \code{eye} Eye where to present the stimulus
\item \code{x} x-coordinate
\item \code{y} y-coordinate
\item \code{sx} Size in the x-axis (diameter for circles).
\item \code{sy} (Optional) Size in the y-axis (diameter for circles). Defaults to sx
\item \code{level} Stimulus level
\item \code{theta} (Optional) Stimulus rotation from 0 to 360 degrees. Defaults to 0
\item \code{t} (Optional) Stimulus presentation time. Defaults to 200 ms
\item \code{window} (Optional) Response window. Defaults to 1500 ms
}}

\item{...}{Parameters for other opiPresent implementations that are ignored here.}
}
\value{
A list containing:
\itemize{
\item \code{err} \code{NULL} if there was no error, a string message if there is an error.
If there was no error the the following fields are also returned.
\item \code{stamp} Time stamp of stimulus onset.
\item \code{time} Response time from stimulus onset if button pressed (ms).
\item \code{seen} \code{TRUE} if seen, \code{FALSE} if not.
\item \code{lpupil}, \code{rpupil} Fixation positions (\code{x}, \code{y}), and diameter (\code{d})
for both left (\code{lpupil}) and right pupils (\code{rpupil}).
}
}
\description{
This is for internal use only. Use \code{\link[=opiPresent]{opiPresent()}} after
\code{chooseOPI("Envision")} to call this function.
}
\examples{
\dontrun{
chooseOpi("Envision")
opiInitialise(address = list(port = 50001, ip = "localhost"))
result <- opiPresent(stim = list(eye = "os", x = 5, y = 4, sx = 0.43, level = 0))
}

}
\seealso{
\code{\link[=opiPresent]{opiPresent()}}
}
