% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNo.r
\name{opiPresent_for_SimNo}
\alias{opiPresent_for_SimNo}
\title{opiPresent_for_SimNo}
\arguments{
\item{stim}{Anything you like, it is ignored.}

\item{...}{Any parameters you like, they are ignored.}
}
\value{
A list with elements:
\itemize{
\item \code{err} Always \code{NULL}.
\item \code{time} Always \code{NA}.
}
}
\description{
Always respond 'not seen' to any parameter.
No checking is done on the validity of \code{stim}.
}
