% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimYes.r
\name{opiPresent_for_SimYes}
\alias{opiPresent_for_SimYes}
\title{opiPresent_for_SimYes}
\arguments{
\item{stim}{Anything you like, it is ignored.}

\item{...}{Any parameters you like, they are ignored.}
}
\value{
A list with elements:
\itemize{
\item \code{err} Always \code{FALSE}.
\item \code{seen} Always \code{TRUE}.
\item \code{time} Always \code{NA}.
}
}
\description{
Always respond 'yes' immediately to any parameter.
}
