% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Envision.r
\name{opiQueryDevice_for_Envision}
\alias{opiQueryDevice_for_Envision}
\title{Implementation of opiQueryDevice for the Envision HMP.}
\value{
A list containing:
\itemize{
\item \code{err} \code{NULL} if there was no error, a string message if there is an error.
\item \code{specs} A list with specifications (see Details) if there was no error.
\item \code{settings} A list with the settings (see Details) if there was no error.
}
}
\description{
This is for internal use only. Use \code{\link[=opiQueryDevice]{opiQueryDevice()}} after
\code{chooseOPI("Envision")} to call this function.
}
\details{
\code{specs} is a list containing:
\itemize{
\item \code{hardware} String with the Hardware name.
\item \code{screens} Whether the hardware has 1 screen (for both eyes) or 2 (1 for each).
\item \code{fov} A list with the field of view for x-axis (\code{x}) and y-axis (\code{y}).
\item \code{resolution} A list with the resolution for width (\code{w}) and height (\code{h}).
\item \code{ipd} A list with min (\code{min}) and max (\code{max}) possible IPDs
and the current value (\code{val}).
\item \code{depth} Display's bit depth.
\item \code{refresh} Display's refresh rate.
\item \code{domain} An array of feasible dB values by the hardware.
}

\code{settings} is a list containing:
\itemize{
\item \code{bgeye} The eye where the background is shown ("OD", "OS", "OU").
\item \code{bglum} Background luminance in cd/m2.
\item \code{fixeye} The eye where a fixation target is shown ("OD", "OS", "OU").
\item \code{fixlum} Fixation luminance in cd/m2.
\item \code{fixtarget} Fixation target.
\item \code{fixcx} Fixation x-coordinate.
\item \code{fixcy} Fixation y-coordinate.
\item \code{fixdx} Fixation size in the x-axis (diameter for circles).
\item \code{fixdy} Fixation size in the y-axis (diameter for circles).
\item \code{fixtheta} Fixation rotation in degrees.
\item \code{tracking} Whether active correction is active.
}
}
\examples{
\dontrun{
chooseOpi("Envision")
opiInitialise(address = list(port = 50001, ip = "localhost"))
opiSetup(list(eye = "BOTH"))
result <- opiQueryDevice()
}

}
\seealso{
\code{\link[=opiQueryDevice]{opiQueryDevice()}}
}
