% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KowaAP7000Client.r
\name{opiQueryDevice_for_KowaAP7000}
\alias{opiQueryDevice_for_KowaAP7000}
\title{Implementation of opiQueryDevice for the Kowa AP7000 machine.}
\value{
List containing
\itemize{
\item \code{isSim = FALSE}.
\item \code{pupilX}, the x-coordinate of the pupil position in pixels.
\item \code{pupilY}, the y-coordinate of the pupil position in pixels.
\item \code{purkinjeX}, the x-coordinate of the purkinje position in pixels.
\item \code{purkinjeY}, the y-coordinate of the purkinje position in pixels.
}
}
\description{
This is for internal use only. Use \code{\link[=opiQueryDevice]{opiQueryDevice()}} with
these Arguments and you will get the Value back.
}
\details{
Prints defined constants in OPI package pertaining to Kowa AP7000.
}
