% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAIA.r
\name{opiQueryDevice_for_MAIA}
\alias{opiQueryDevice_for_MAIA}
\title{Implementation of opiQueryDevice for the ImoVifa machine.}
\value{
A list containing constants and their values used in the OPI MAIA module.
}
\description{
This is for internal use only. Use \code{\link[=opiQueryDevice]{opiQueryDevice()}} with
these Arguments and you will get the Value back.
}
