% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Octopus600.r
\name{opiQueryDevice_for_O600}
\alias{opiQueryDevice_for_O600}
\title{Implementation of opiQueryDevice for the O600 machine.}
\value{
Returns a list of 10 items:

\if{html}{\out{<div class="sourceCode">}}\preformatted{* \code{answerButton} 0 = not pressed, 1 = pressed
* \code{headSensor} 0 = no forehead detected, 1 = forehead detected
* \code{eyeLidClosureLeft} 0 = eye is open, 1 = eye is closed
* \code{eyeLidClosureRight} 0 = eye is open, 1 = eye is closed
* \code{fixationLostLeft} 1 = eye pos lost, 0 = eye pos ok
* \code{fixationLostRight} 1 = eye pos lost, 0 = eye pos ok
* \code{pupilPositionXLeft} (in px)
* \code{pupilPositionYLeft} (in px)
* \code{pupilPositionXRight} (in px)
* \code{pupilPositionYRight} (in px)
}\if{html}{\out{</div>}}
}
\description{
This is for internal use only. Use \code{\link[=opiQueryDevice]{opiQueryDevice()}} with
these Arguments and you will get the Value back.
}
\details{
Prints defined constants in OPI package pertaining to O600.
}
