% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Octopus900.r
\name{opiQueryDevice_for_Octopus900}
\alias{opiQueryDevice_for_Octopus900}
\title{Implementation of opiQueryDevice for the Octopus900 machine.}
\value{
List containing \code{isSim = FALSE}.
}
\description{
This is for internal use only. Use \code{\link[=opiQueryDevice]{opiQueryDevice()}} with
these Arguments and you will get the Value back.
}
\details{
Prints defined constants in OPI package pertaining to Octopus 900.
}
