% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PhoneHMD.r
\name{opiQueryDevice_for_PhoneHMD}
\alias{opiQueryDevice_for_PhoneHMD}
\title{Implementation of opiQueryDevice for the PhoneHMD machine.}
\value{
A list containing:
\itemize{
\item \code{err} \code{NULL} if there was no error, a string message if there is an error.
}
}
\description{
This is for internal use only. Use \code{\link[=opiQueryDevice]{opiQueryDevice()}} after
\code{chooseOPI("PhoneHMD")} to call this function.
}
\examples{
chooseOpi("PhoneHMD")
opiInitialise(list(port = 50001, ip = "localhost"))
opiSetup(list(eye = "BOTH"))
result <- opiQueryDevice()

}
\seealso{
\code{\link[=opiQueryDevice]{opiQueryDevice()}}
}
