% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimGaussian.r
\name{opiQueryDevice_for_SimGaussian}
\alias{opiQueryDevice_for_SimGaussian}
\title{opiQueryDevice_for_SimGaussian}
\value{
A list with elements:
\itemize{
\item \code{err} Always \code{NULL}
\item \code{machine} that is set to \code{"SimGaussian"}.
}
}
\description{
Returns a simple list.
}
