% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimHenson.r
\name{opiQueryDevice_for_SimHenson}
\alias{opiQueryDevice_for_SimHenson}
\title{opiQueryDevice_for_SimHenson}
\value{
A list with elements:
\itemize{
\item \code{isSim} Always \code{TRUE}.
\item \code{machine} that is set to \code{"SimHenson"}.
}
}
\description{
Returns name of the machine.
}
