% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimHensonRT.r
\name{opiQueryDevice_for_SimHensonRT}
\alias{opiQueryDevice_for_SimHensonRT}
\title{opiQueryDevice_for_SimHensonRT}
\value{
A list with elements:
\itemize{
\item \code{isSim} Always \code{TRUE}.
\item \code{machine} that is set to \code{"SimHensonRT"}.
}
}
\description{
Returns name of the machine.
}
