% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNo.r
\name{opiQueryDevice_for_SimNo}
\alias{opiQueryDevice_for_SimNo}
\title{opiQueryDevice_for_SimNo}
\value{
A list with elements:
\itemize{
\item \code{machine} that is set to \code{"SimNo"}.
\item \code{isSim} that is set to TRUE.
}
}
\description{
Returns name of the machine.
}
