% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimYes.r
\name{opiQueryDevice_for_SimYes}
\alias{opiQueryDevice_for_SimYes}
\title{opiQueryDevice_for_SimYes}
\value{
A list with elements:
\itemize{
\item \code{isSim} Always \code{TRUE}.
\item \code{machine} that is set to \code{"SimYes"}.
}
}
\description{
Returns name of the machine.
}
