% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi.r
\name{opiSetup}
\alias{opiSetup}
\title{Calls opiSetup_for_MACHINE as appropriate.}
\usage{
opiSetup(settings)
}
\arguments{
\item{settings}{A list containing the same names as that returned by \link{opi_queryDevice}.}
}
\value{
Each implementation should(!) return a list with at least the following elements:
\itemize{
\item \code{err} \code{NULL} if no error, otherwise a string describing the error.
}
}
\description{
Specific parameters and return values can be seen in the machine specific versions
listed below in the ’See Also’.
}
\seealso{
\code{\link[=opiSetup_for_Compass]{opiSetup_for_Compass()}},
\code{\link[=opiSetup_for_Octopus900]{opiSetup_for_Octopus900()}},
\code{\link[=opiSetup_for_ImoVifa]{opiSetup_for_ImoVifa()}},
\code{\link[=opiSetup_for_PhoneHMD]{opiSetup_for_PhoneHMD()}}, \code{\link[=opiSetup_for_Display]{opiSetup_for_Display()}}, \code{\link[=opiSetup_for_PicoVR]{opiSetup_for_PicoVR()}},
\code{\link[=opiSetup_for_SimNo]{opiSetup_for_SimNo()}}, \code{\link[=opiSetup_for_SimYes]{opiSetup_for_SimYes()}}, \code{\link[=opiSetup_for_SimHenson]{opiSetup_for_SimHenson()}},
\code{\link[=opiSetup_for_SimGaussian]{opiSetup_for_SimGaussian()}}
}
