% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Envision.r
\name{opiSetup_for_Envision}
\alias{opiSetup_for_Envision}
\title{Implementation of opiSetup for the Envision HMP.}
\arguments{
\item{settings}{A list containing:
\itemize{
\item \code{bgeye} The eye where the background is shown ("OD", "OS", "OU"). Default is "OU".
\item \code{bglum} Background luminance in cd/m2. Default is 10.
\item \code{fixeye} The eye where a fixation target is shown ("OD", "OS", "OU"). Default is "OU".
\item \code{fixlum} Fixation luminance in cd/m2. Default is 20.
\item \code{fixtarget} Fixation target. Default is "CROSS".
\item \code{fixcx} Fixation x-coordinate. Default is 0.
\item \code{fixcy} Fixation y-coordinate. Default is 0.
\item \code{fixdx} Fixation size in the x-axis (diameter for circles). Default is 2.
\item \code{fixdy} Fixation size in the y-axis (diameter for circles). Default is 2.
\item \code{fixtheta} Fixation rotation in degrees. Default is 0.
\item \code{tracking} Whether active correction is active. Default is False.
}}
}
\value{
A list containing:
\itemize{
\item \code{err} \code{NULL} if there was no error, a string message if there is an error.
}
}
\description{
This is for internal use only. Use \code{\link[=opiSetup]{opiSetup()}} after
\code{chooseOPI("Envision")} to call this function.
}
\details{
All fields are optional. Missing fields will be assigned their Default value.
\code{opiSetup()} sets all to default.
}
\examples{
\dontrun{
chooseOpi("Envision")
opiInitialise(address = list(port = 50001, ip = "localhost"))
result <- opiSetup(settings = list(eye = "BOTH"))
}

}
\seealso{
\code{\link[=opiSetup]{opiSetup()}}
}
