% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Octopus600.r
\name{opiSetup_for_O600}
\alias{opiSetup_for_O600}
\title{opiSetup for the O600 machine.}
\arguments{
\item{bgColor}{Background color 0 to 255.}

\item{fixType}{fixation type 1, 2, 3 or 4.}

\item{fixColor}{fixation color 0, 2, 4 or 5.}

\item{fixIntensity}{fixation point intensity 0 to 255.}
}
\value{
A list containing one of the following
\itemize{
\item \code{err = -1} to be implemented
\item \code{err = -2} O600 sent back an error; bad background parameters
\item \code{err = -3} O600 sent back an error; bad fixation parameters
\item \code{err = NULL} Success
}
}
\description{
Implementation of opiSetup for the O600 machine.

This is for internal use only. Use \code{\link[=opiSetup]{opiSetup()}} with these Arguments.
}
