% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimGaussian.r
\name{opiSetup_for_SimGaussian}
\alias{opiSetup_for_SimGaussian}
\title{opiSetup_for_SimGaussian}
\arguments{
\item{...}{Any object you like, it is ignored.}
}
\value{
A list with elements:
\itemize{
\item \code{err} Always \code{NULL}.
}
}
\description{
Does nothing.
}
