% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimHensonRT.r
\name{opiSetup_for_SimHensonRT}
\alias{opiSetup_for_SimHensonRT}
\title{opiSetup_for_SimHensonRT}
\arguments{
\item{...}{Any object you like, it is ignored.}
}
\value{
A list with elements:
\itemize{
\item \code{err} Always \code{NULL}.
}
}
\description{
Does nothing.
For internal use only, use \code{opiSetup()}.
}
