% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimNo.r
\name{opiSetup_for_SimNo}
\alias{opiSetup_for_SimNo}
\title{opiSetup_for_SimNo}
\arguments{
\item{settings}{Anything you like, it is ignored.}
}
\value{
A list with elements:
\itemize{
\item \code{err} Always \code{NULL}.
}
}
\description{
Does nothing.
}
