% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimYes.r
\name{opiSetup_for_SimYes}
\alias{opiSetup_for_SimYes}
\title{opiSetup_for_SimYes}
\arguments{
\item{settings}{Any object you like, it is ignored.}
}
\value{
A list with elements:
\itemize{
\item \code{err} Always \code{NULL}.
}
}
\description{
Does nothing.
}
