% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pix2deg.r
\name{pixToDeg}
\alias{pixToDeg}
\title{convert pixels to degrees for machine 'machine'}
\usage{
pixToDeg(xy, machine = "compass")
}
\arguments{
\item{xy}{a 2 element vector c(x,y) where x and y are in pixels}

\item{machine}{"compass" or ...?}
}
\value{
xy converted to degrees of visual field with the usual conventions or \code{NA} if machine is unknown
}
\description{
convert pixels to degrees for machine 'machine'
}
\examples{
pixToDeg(c(1000, 200), machine="compass")  # c(1.290323, 24.516129) degrees
pixToDeg(c(1920/2, 1920/2)) # c(0,0) degrees

}
