% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-RtSigmaUnits.r
\docType{data}
\name{RtSigmaUnits}
\alias{RtSigmaUnits}
\title{Response times to white-on-white Goldmann Size III targets for 12 subjects
in sigma units}
\format{
An object of class \code{data.frame} with 30620 rows and 3 columns.
}
\usage{
RtSigmaUnits
}
\description{
Response times to white-on-white Goldmann Size III targets for 12
subjects. The second column is the distance of the stimuli from measured threshold
in 'sigma' units. The threshold was determined by post-hoc fit of a cummulative
gaussian FoS curve to the data for each location and subject. Sigma is the
standard deviation of the fitted FoS.
}
\details{
A data frame with 30620 observations on the following 3 variables.
\describe{
  \item{\code{Rt}}{Reaction time in ms.}
  \item{\code{Dist}}{Distance of stimuli from threshold in sigma units.}
  \item{\code{Person}}{Identifier of each subject.}
}
}
\references{
A.M. McKendrick, J. Denniss and A. Turpin. "Response times across the visual
field: empirical observations and application to threshold determination".
Vision Research 101 2014.
}
\seealso{
\code{\link{RtDbUnits}}
}
\keyword{dataset}
