\name{OPI-package}
\alias{OPI-package}
\alias{OPI}
\docType{package}
\title{
Open Perimetry Interface
}
\description{

This package contains the current version of the Open Perimetry Interface (OPI), for controlling visual field machines using R. The OPI is a standard for interfacing with visual field testing machines (perimeters). It specifies basic functions that allow many visual field tests to be constructed. As of July 2014 it is fully implemented on the Octopus 900 and Octopus 600, and partially on the Heidelberg Edge Perimeter. For the HEP, you will need the Rhep package and permission from Heidelberg Engineering. For the Octopus 900, you will need the O900Server.jar program and for both Octopus machines you will need permission from Haag-Streit. See the OPI www site for details. There is also a cousin: the R package visualFields, which has tools for analysing and manipulating visual field data.

}
\details{
    \tabular{ll}{
     Package: \tab OPI\cr
     Type: \tab Package\cr
     Version: \tab 1.4\cr
     Date: \tab 2012-10-26\cr
     License: \tab GPL-3\cr
    }
}
\author{
Andrew Turpin <aturpin@unimelb.edu.au>
Maintainer: Andrew Turpin <aturpin@unimelb.edu.au>
}
\references{
Please cite:
A. Turpin, P.H. Artes and A.M. McKendrick
"The Open Perimetry Interface: An enabling tool for clinical visual psychophysics",
Journal of Vision 12(11) 2012.

http://perimetry.org/OPI
}
\keyword{ package }
\seealso{
visualFields
}
