% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OPI-package.r
\docType{package}
\name{OPI-package}
\alias{OPI}
\alias{OPI-package}
\title{OPI: Open Perimetry Interface}
\description{
Implementation of the Open Perimetry Interface (OPI) for simulating and controlling visual field machines using R. The OPI is a standard for interfacing with visual field testing machines (perimeters) first started as an open source project with support of Haag-Streit in 2010. It specifies basic functions that allow many visual field tests to be constructed. As of February 2022 it is fully implemented on the Haag-Streit Octopus 900 with partial implementations on the Centervue Compass, Kowa AP 7000, Android phones and the CrewT IMO. It also has a cousin: the R package 'visualFields', which has tools for analysing and manipulating visual field data.
}
\note{
`.OpiEnv$machine_name` is set to the name of the chosen OPI set by the `chooseOPI` function call.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://people.eng.unimelb.edu.au/aturpin/opi/index.html}
}

}
\author{
\strong{Maintainer}: Andrew Turpin \email{andrew.turpin@lei.org.au} [copyright holder]

Other contributors:
\itemize{
  \item David Lawson [contributor, copyright holder]
  \item Matthias Muller [contributor]
  \item Jonathan Dennis [contributor, copyright holder]
  \item Astrid Zeman [contributor]
  \item Ivan Marin-Franch [contributor]
}

}
\keyword{internal}
