\name{OPI-package}
\alias{OPI-package}
\alias{OPI}
\docType{package}
\title{
Open Perimetry Interface
}
\description{

Implementation of the Open Perimetry Interface (OPI) for simulating and controlling visual field machines using R. The OPI is a standard for interfacing with visual field testing machines (perimeters). It specifies basic functions that allow many visual field tests to be constructed. As of October 2017 it is fully implemented on the Octopus 900 and partially on the Heidelberg Edge Perimeter, the Kowa AP 7000, the CrewT imo and the Centervue Compass. It also has a cousin: the R package 'visualFields', which has tools for analysing and manipulating visual field data.
License: GPL-3

}
\details{
    \tabular{ll}{
     Package: \tab OPI\cr
     Type: \tab Package\cr
     Version: \tab 2.9\cr
     Date: \tab 2012-10-26\cr
     License: \tab GPL-3\cr
    }
}
\author{
Andrew Turpin <aturpin@unimelb.edu.au>
Maintainer: Andrew Turpin <aturpin@unimelb.edu.au>
}
\references{
Please cite:
A. Turpin, P.H. Artes and A.M. McKendrick
"The Open Perimetry Interface: An enabling tool for clinical visual psychophysics",
Journal of Vision 12(11) 2012.

http://perimetry.org/OPI
}
\keyword{ package }
\seealso{
visualFields
}
