% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi.r, R/Octopus600.r
\name{opiClose}
\alias{opiClose}
\alias{octo600.opiClose}
\title{Calls opiClose_for_MACHINE as appropriate.}
\usage{
opiClose()

octo600.opiClose()
}
\value{
Each implementation should(!) return a list with at least the following elements:
\itemize{
\item \code{err} \code{NULL} if no error, otherwise a string describing the error.
}

\subsection{Octopus600}{

Always returns NULL
}
}
\description{
Specific parameters and return values can be seen in the machine specific versions
listed below in the ’See Also’.
}
\seealso{
\code{\link[=opiClose_for_Compass]{opiClose_for_Compass()}},
\code{\link[=opiClose_for_Octopus900]{opiClose_for_Octopus900()}},
\code{\link[=opiClose_for_ImoVifa]{opiClose_for_ImoVifa()}},
\code{\link[=opiClose_for_PhoneHMD]{opiClose_for_PhoneHMD()}}, \code{\link[=opiClose_for_Display]{opiClose_for_Display()}}, \code{\link[=opiClose_for_PicoVR]{opiClose_for_PicoVR()}},
\code{\link[=opiClose_for_SimNo]{opiClose_for_SimNo()}}, \code{\link[=opiClose_for_SimYes]{opiClose_for_SimYes()}}, \code{\link[=opiClose_for_SimHenson]{opiClose_for_SimHenson()}},
\code{\link[=opiClose_for_SimGaussian]{opiClose_for_SimGaussian()}}
}
