% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi.r
\name{opiClose}
\alias{opiClose}
\title{Close using OPI}
\usage{
opiClose(...)
}
\arguments{
\item{...}{Implementation specific parameters. See details.}
}
\value{
Returns NULL if close succeeded, otherwise an implementation-dependent
error.
\subsection{Compass}{
  Returns a list of \code{err}, which is an error code, and \code{fixations},
  which is a matrix with three columns: time (same as \code{time_hw}
  in \code{opiPresent}), x (degrees relative to the centre of the image
  returned by \code{opiInitialise} - not the PRL), y (as for x), and one row
  per fixation.
}
}
\description{
Generic function for closing the chosen OPI implementation that is set
with \code{chooseOpi()}
}
\examples{
chooseOpi("SimGaussian")
if (!is.null(opiInitialize(sd=2)))
  stop("opiInitialize failed")
if (!is.null(opiClose()))
  stop("opiClose failed, which is very surprising!")
}
\seealso{
\code{\link{chooseOpi}}
}
