% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compass.r
\name{opiClose_for_Compass}
\alias{opiClose_for_Compass}
\title{Implementation of opiClose for the Compass machine.}
\value{
A list with elements
\itemize{
\item \code{err}, which is an error code, NULL for no error
\item \code{fixations}, which is a matrix one row per fixation and three columns:
\itemize{
\item \code{time} (same as \code{time_hw} in \code{opiPresent})
\item \code{x} (degrees relative to the centre of the image returned by \code{opiInitialise} - not the PRL)
\item \code{y} (as for x)
}
}
}
\description{
This is for internal use only. Use \code{\link[=opiClose]{opiClose()}} with
these Arguments and you will get the Value back.
}
