% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi.r, R/Octopus600.r
\name{opiQueryDevice}
\alias{opiQueryDevice}
\alias{octo600.opiQueryDevice}
\title{Calls opiQueryDevice_for_MACHINE as appropriate.}
\usage{
opiQueryDevice()

octo600.opiQueryDevice()
}
\value{
A list specific to each machine.

\subsection{Octopus600}{

Returns a list of 10 items:
* \code{answerButton} \{0 = not pressed, 1 = pressed \}
* \code{headSensor} \{0 = no forehead detected, 1 = forehead detected \}
* \code{eyeLidClosureLeft} \{0 = eye is open, 1 = eye is closed\}
* \code{eyeLidClosureRight} \{0 = eye is open, 1 = eye is closed\}
* \code{fixationLostLeft} \{1 = eye pos lost, 0 = eye pos ok\}
* \code{fixationLostRight} \{1 = eye pos lost, 0 = eye pos ok\}
* \code{pupilPositionXLeft} (in px)
* \code{pupilPositionYLeft} (in px)
* \code{pupilPositionXRight} (in px)
* \code{pupilPositionYRight} (in px)
}
}
\description{
Returns a list that describes the current state of the machine.
Specific parameters and return values can be seen in the machine specific versions
listed below in the ’See Also’.
}
\section{Octopus600}{
If the chosen OPI is \code{Octopus600}, then this function returns
information about the patient. See the Value section for details.
}

\seealso{
\code{\link[=opiQueryDevice_for_ImoVifa]{opiQueryDevice_for_ImoVifa()}},
\code{\link[=opiQueryDevice_for_Compass]{opiQueryDevice_for_Compass()}},
\code{\link[=opiQueryDevice_for_Octopus900]{opiQueryDevice_for_Octopus900()}},
\code{\link[=opiQueryDevice_for_PhoneHMD]{opiQueryDevice_for_PhoneHMD()}}, \code{\link[=opiQueryDevice_for_Display]{opiQueryDevice_for_Display()}}, \code{\link[=opiQueryDevice_for_PicoVR]{opiQueryDevice_for_PicoVR()}},
\code{\link[=opiQueryDevice_for_SimNo]{opiQueryDevice_for_SimNo()}}, \code{\link[=opiQueryDevice_for_SimYes]{opiQueryDevice_for_SimYes()}}, \code{\link[=opiQueryDevice_for_SimHenson]{opiQueryDevice_for_SimHenson()}},
\code{\link[=opiQueryDevice_for_SimGaussian]{opiQueryDevice_for_SimGaussian()}}
}
