% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi.r
\name{opiQueryDevice}
\alias{opiQueryDevice}
\title{Calls opiQueryDevice_for_MACHINE as appropriate.}
\usage{
opiQueryDevice()
}
\value{
A list specific to each machine.
}
\description{
Returns a list that describes the current state of the machine.
Specific parameters and return values can be seen in the machine specific versions
listed below in the ’See Also’.
}
\seealso{
\code{\link[=opiQueryDevice_for_ImoVifa]{opiQueryDevice_for_ImoVifa()}},
\code{\link[=opiQueryDevice_for_Compass]{opiQueryDevice_for_Compass()}},
\code{\link[=opiQueryDevice_for_Octopus900]{opiQueryDevice_for_Octopus900()}},
\code{\link[=opiQueryDevice_for_PhoneHMD]{opiQueryDevice_for_PhoneHMD()}}, \code{\link[=opiQueryDevice_for_Display]{opiQueryDevice_for_Display()}}, \code{\link[=opiQueryDevice_for_PicoVR]{opiQueryDevice_for_PicoVR()}},
\code{\link[=opiQueryDevice_for_SimNo]{opiQueryDevice_for_SimNo()}}, \code{\link[=opiQueryDevice_for_SimYes]{opiQueryDevice_for_SimYes()}}, \code{\link[=opiQueryDevice_for_SimHenson]{opiQueryDevice_for_SimHenson()}},
\code{\link[=opiQueryDevice_for_SimGaussian]{opiQueryDevice_for_SimGaussian()}}
}
