% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi.r, R/Octopus600.r
\name{opiSetBackground}
\alias{opiSetBackground}
\alias{octo600.opiSetBackground}
\title{Deprecated. Use \code{\link[=opiSetup]{opiSetup()}}.}
\usage{
octo600.opiSetBackground(
  bgColor = NA,
  fixType = NA,
  fixColor = NA,
  fixIntensity = 255
)
}
\arguments{
\item{bgColor}{Background color}

\item{fixType}{fixation type}

\item{fixColor}{fixation color}
}
\value{
\subsection{Octopus600}{
\itemize{
\item -1 to be implemented
\item -2 O600 sent back an error; bad background parameters
\item -3 O600 sent back an error; bad fixation parameters
\item NULL Success
}
}
}
\description{
In older OPIs it set background color and luminance in both eyes.
Deprecated for OPI >= v3.0.0 and replaced with \code{\link[=opiSetup]{opiSetup()}}.
}
\section{Octopus600}{
This function has no effect.
}

\seealso{
\code{\link[=opiSetup]{opiSetup()}}
}
