% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi.r
\name{opiSetBackground}
\alias{opiSetBackground}
\title{Deprecated. Use \code{\link[=opiSetup]{opiSetup()}}.}
\description{
In older OPIs it set background color and luminance in both eyes.
Deprecated for OPI >= v3.0.0 and replaced with \code{\link[=opiSetup]{opiSetup()}}.
}
\seealso{
\code{\link[=opiSetup]{opiSetup()}}
}
