% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi.r
\name{opiStaticStimulus}
\alias{opiStaticStimulus}
\title{For backwards compatibility. Used by Octopus900 and KowaAP7000.}
\usage{
opiStaticStimulus()
}
\description{
For backwards compatibility. Used by Octopus900 and KowaAP7000.
}
