% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opi.r
\name{opiTemporalStimulus}
\alias{opiTemporalStimulus}
\title{For backwards compatibility. Used by Octopus900 and KowaAP7000.}
\usage{
opiTemporalStimulus()
}
\description{
For backwards compatibility. Used by Octopus900 and KowaAP7000.
}
