% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-RtDbUnits.r
\docType{data}
\name{RtDbUnits}
\alias{RtDbUnits}
\title{Response times to white-on-white Goldmann Size III targets for 12 subjects
in dB units}
\format{
An object of class \code{data.frame} with 30620 rows and 3 columns.
}
\usage{
RtDbUnits
}
\description{
Response times to white-on-white Goldmann Size III targets for 12
subjects. The second column is the distance of the stimuli from measured
threshold in HFA dB units. The threshold was determined by post-hoc fit of FoS
curves to the data.
}
\details{
A data frame with 30620 observations on the following 3 variables.
\itemize{
\item \code{Rt} Reaction time in ms.
\item \code{Dist} Distance of stimuli from threshold in dB.
\item \code{Person} Identifier of each subject.
}
}
\references{
A.M. McKendrick, J. Denniss and A. Turpin. "Response times across the visual
field: empirical observations and application to threshold determination".
Vision Research 101 2014.
}
\seealso{
\code{\link{RtSigmaUnits}}
}
\keyword{dataset}
